# public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Example API",
        "description": "Example info",
        "contact": {
            "name": "Swagger API Team"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "https://example.localhost",
            "description": "API server"
        }
    ],
    "paths": {
...
        "/auth/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register a new user",
                "description": "Registers a new user. Optionally creates a new troop or joins an existing patrol via invite code.",
                "operationId": "10c02a951e844462c465d0e576b8c942",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nickname",
                                    "name",
                                    "surname",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "nickname": {
                                        "type": "string",
                                        "example": "scoutjoe"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Joe"
                                    },
                                    "surname": {
                                        "type": "string",
                                        "example": "Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "joe@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "hunter2"
                                    },
                                    "new_troop": {
                                        "properties": {
                                            "name": {
                                                "type": "string",
                                                "example": "Falcons"
                                            }
                                        },
                                        "type": "object"
                                    },
                                    "invite_code": {
                                        "type": "string",
                                        "example": "a1b2c3d4"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User successfully registered"
                    },
                    "400": {
                        "description": "Missing required troop info"
                    },
                    "409": {
                        "description": "Email already exists"
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Log in a user",
                "description": "Logs a user in by verifying email and password.\nReturns access token, refresh token and user with roles.",
                "operationId": "363e422b5a8c763fa1ea6fd03a77481d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "joe@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "hunter2"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successfully logged in"
                    },
                    "400": {
                        "description": "Missing credentials"
                    },
                    "401": {
                        "description": "Invalid login"
                    }
                }
            }
        },
        "/auth/refresh": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Refresh access token using refresh token",
                "description": "Issues a new access token based on a valid refresh token.",
                "operationId": "d6e134316a3213ab2da85eb1e545a77f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "refresh_token"
                                ],
                                "properties": {
                                    "refresh_token": {
                                        "type": "string",
                                        "example": "123456789abcdef"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Access token refreshed"
                    },
                    "400": {
                        "description": "Missing refresh token"
                    },
                    "401": {
                        "description": "Invalid or expired refresh token"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            }
        },
...
    },
...
}